+++ nodalia-cards-2.1.0.js
@@ -1,2 +1,0 @@
-import "./nodalia-cards-2.1.2-alpha.1.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.2-alpha.1","contentSha256_12":"c16875640cc5","file":"nodalia-cards-2.1.0.js","targetFile":"nodalia-cards-2.1.2-alpha.1.js","fallbackFile":"nodalia-cards.js"};}